ajout dislike

+++ php/like.php
@@ -13,6 +13,14 @@ mysql_query("SET NAMES UTF8");
 			mysql_query(" UPDATE ARTICLE SET nb_like = nb_like + 1 WHERE id_art = ".$_GET['id_art'].";");
 			
 			mysql_query("INSERT INTO ARTICLE_FAV(login,id_art) VALUES('".$_GET['login']."',".$_GET['id_art'].");");
+			
+			$query = mysql_query("SELECT * FROM ARTICLE_DISLIKE where login = '".$_GET['login']."';");
+			
+			if(mysql_num_rows($query)!=0){
+			
+				mysql_query("DELETE FROM ARTICLE_DISLIKE WHERE login = '".$_GET['login']."' and id_art = '".$_GET['id_art']."' ;");
+				
+			}
 		}
 		else{
 			echo "dejalike";
@@ -21,15 +29,27 @@ mysql_query("SET NAMES UTF8");
 	}
 	else if($_GET['type']=="dislike"){
 		
-		mysql_query(" UPDATE ARTICLE SET nb_like = nb_like - 1 WHERE id_art = ".$_GET['id_art'].";");
+		$query_dejadislike = mysql_query("SELECT * FROM ARTICLE_DISLIKE WHERE login = '".$_GET['login']."' and id_art = ".$_GET['id_art'].";");
 		
-		$query = mysql_query("SELECT * FROM ARTICLE_FAV where login = '".$_GET['login']."';");
+		if(mysql_num_rows($query_dejadislike)==0){
 		
-		if(mysql_num_rows($query)!=0){
-		
-			mysql_query("DELETE FROM ARTICLE_FAV WHERE login = '".$_GET['login']."' and id_art = '".$_GET['id_art']."' ;");
+			mysql_query("UPDATE ARTICLE SET nb_like = nb_like - 1 WHERE id_art = ".$_GET['id_art'].";");
 			
+			mysql_query("INSERT INTO ARTICLE_DISLIKE(login,id_art) VALUES('".$_GET['login']."',".$_GET['id_art'].");");
+			
+			$query = mysql_query("SELECT * FROM ARTICLE_FAV where login = '".$_GET['login']."';");
+			
+			if(mysql_num_rows($query)!=0){
+			
+				mysql_query("DELETE FROM ARTICLE_FAV WHERE login = '".$_GET['login']."' and id_art = '".$_GET['id_art']."' ;");
+				
+			}
+		}
+		else{
+		
+			echo "dejadislike";
+				
+			}
 		}
 	
-	}
 	?>
